+++ claroil/well-opportunity-c-sharp/application/logic/Schemas/QueryObject.cs
@@ -151,7 +151,7 @@ namespace netBox.Schemas
           }
         ),
         resolve: context => wellRepository.ShouldTestWell(
-          context.GetArgument<float>("healthIndex"),
+          context.GetArgument<double>("healthIndex"),
           context.GetArgument<int>("lastTestDay"),
           context.GetArgument<int>("today"),
           context.CancellationToken
@@ -281,7 +281,7 @@ namespace netBox.Schemas
           }
         ),
         resolve: context => wellRepository.InterventionRevenueGain(
-          context.GetArgument<float>("oilPrice"),
+          context.GetArgument<double>("oilPrice"),
           context.GetArgument<Metrics>("measuredMetrics"),
           context.GetArgument<ActionOutcome>("revenueGains"),
           context.CancellationToken
@@ -309,7 +309,7 @@ namespace netBox.Schemas
           }
         ),
         resolve: context => wellRepository.SkippingTestCostReduction(
-          context.GetArgument<float>("oilPrice"),
+          context.GetArgument<double>("oilPrice"),
           context.GetArgument<Metrics>("measuredMetrics"),
           context.GetArgument<ActionOutcome>("revenueGains"),
           context.CancellationToken
